+++ src/main/java/org/eurostates/area/ESUser.java
@@ -1,6 +1,6 @@
 package org.eurostates.area;
 
-import org.eurostates.area.Rank;
+import org.eurostates.area.state.CustomState;
 import org.eurostates.area.state.State;
 import org.eurostates.area.state.States;
 import org.eurostates.area.town.Town;
@@ -23,18 +23,21 @@ public class ESUser {
     }
 
     public Optional<Rank> getRank() {
-        Optional<Town> opTown = getTown();
-        return opTown.flatMap(town -> town
-                .getRanks()
+        State state = getState();
+        return state.getRanks()
                 .parallelStream()
                 .filter(rank -> rank.getId().equals(this.rankUUID))
-                .findAny());
+                .findAny();
 
     }
 
     public void setRank(Rank rank) {
-        Town town = getTown().orElseThrow(() -> new IllegalStateException("Player must be part of a town before setting a rank"));
-        if (!rank.getState().equals(town.getState())) {
+        State state = getState();
+        if (!(state instanceof CustomState)) {
+            throw new IllegalStateException("Player must be part of a town before setting a rank");
+        }
+
+        if (!rank.getState().equals(state)) {
             throw new IllegalStateException("Rank is not part of the players state");
         }
         this.rankUUID = rank.getId();
@@ -45,7 +48,7 @@ public class ESUser {
                 .CUSTOM_STATES
                 .parallelStream()
                 .flatMap(state -> state.getTowns().stream())
-                .filter(town -> town.getCitizenIds().contains(this.uuid))
+                .filter(town -> town.getOwnerId().equals(this.uuid))
                 .findAny();
     }
